+++ htdocs/pages/wiki.php
@@ -2,7 +2,15 @@
 <html>
 
 <body>
-<a href="https://matchbox.fandom.com/wiki/Matchbox_Cars_Wiki">Front Page</a>
+<ul>
+<li><a href="https://matchbox.fandom.com/wiki/Matchbox_Cars_Wiki">Front Page</a>
+<li><a href="https://matchbox.fandom.com/wiki/Models_of_Yesteryear">Models of Yesteryear</a>
+<li><a href="https://matchbox.fandom.com/wiki/Sky_Busters">Skybusters</a>
+<li><a href="https://matchbox.fandom.com/wiki/Speed_Kings">Speed Kings</a>
+<li><a href="https://matchbox.fandom.com/wiki/Super_Kings">Super Kings</a>
+<li><a href="https://matchbox.fandom.com/wiki/Sea_Kings">Sea Kings</a>
+<li><a href="https://matchbox.fandom.com/wiki/Battle_Kings">Battle Kings</a>
+</ul><hr>
 
 <table>
 <tr style='vertical-align: top;'>